Ticket 232

+++ sic-api/src/main/java/sic/modelo/BusquedaFacturaVentaCriteria.java
@@ -22,8 +22,8 @@ public class BusquedaFacturaVentaCriteria {
     private boolean buscaUsuario;
     private Usuario usuario;
     private boolean buscaPorNumeroFactura;
-    private int numSerie;
-    private int numFactura;
+    private long numSerie;
+    private long numFactura;
     private boolean buscarPorPedido;
     private long nroPedido;
     private boolean buscaSoloImpagas;
